SSBDM/requirements.txt
Django>=5.1.7
psycopg2>=2.9.9
psycopg2-binary>=2.9.9
django-mptt>=0.16.0
django-filter==23.5
Pillow>=10.3.0
django-crispy-forms>=2.1
crispy-bootstrap5>=2023.10
djangorestframework>=3.14.0
django-widget-tweaks==1.5.0
python-dateutil==2.8.2
whitenoise==6.6.0
markdown==3.5.2
django-tables2==2.7.0
django-import-export==3.3.6
django-debug-toolbar>=4.3.0
python-dotenv>=1.0.0